separated backend code into modules
